external-mvl-toolchain: Handle multilib arches with different package arches
Source: MontaVista Software, LLC
MR: 72632
Type: Integration
Disposition: Local
ChangeID: 48b313320600366aee6c69d7f67959d37db21ece
Description:

+++ recipes-mvista/meta/external-mvl-toolchain.bb
@@ -196,7 +196,13 @@ do_install() {
      if [ "${TARGET_VENDOR_MULTILIB_ORIGINAL}" != "" -a "${TARGET_VENDOR}" != "${TARGET_VENDOR_MULTILIB_ORIGINAL}" ]; then
        for each in ${D}/${includedir}/glibc-locale-internal-*; do 
          if [ "x$each" != "x${D}/${includedir}/glibc-locale-internal-${MULTIMACH_TARGET_SYS}" ] ; then
-               rm -rf $each
+               if [ -z "$(basename $each | grep mllib)" ] ; then
+                  rm -rf $each
+               else
+                  if [ ! -d ${D}/${includedir}/glibc-locale-internal-${MULTIMACH_TARGET_SYS} ] ; then
+                     mv $each ${D}/${includedir}/glibc-locale-internal-${MULTIMACH_TARGET_SYS}
+                  fi
+               fi
          fi
        done
      else
